phase_2_active/langchain_fundamentals/week2_day14-16/requirements.txt
# 🎯 Week 2: Days 14-16 - Sequential Chains & Memory
# Complete dependency list for LangChain with memory systems

# =========== CORE LANGCHAIN ===========
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10

# =========== AI/ML ===========
openai==1.3.0
tiktoken==0.5.2                    # Token counting
sentence-transformers==2.2.2       # Local embeddings (optional)

# =========== VECTOR DATABASES ===========
chromadb==0.4.18                   # Vector storage for memory
pypdf==3.17.4                      # PDF processing (for document memory)
unstructured==0.12.2               # Document parsing

# =========== DATA & UTILITIES ===========
python-dotenv==1.0.0              # Environment variables
pydantic==2.5.0                   # Data validation
orjson==3.9.10                    # Fast JSON handling
redis==5.0.1                      # In-memory data store (for memory caching)

# =========== TERMINAL UI ===========
rich==13.7.0                      # Beautiful terminal output
colorama==0.4.6                   # Cross-platform colored text
tqdm==4.66.1                      # Progress bars

# =========== DEVELOPMENT ===========
pytest==7.4.3                     # Testing framework
black==23.11.0                    # Code formatting
flake8==6.1.0                     # Code linting
mypy==1.7.0                       # Type checking

# =========== OPTIONAL (for advanced memory) ===========
# Uncomment if needed:
# faiss-cpu==1.7.4                # Vector similarity search
# sentencepiece==0.1.99           # Tokenization for local models
# transformers==4.36.2            # Hugging Face models